Reorganize tests into unit and integration directories
Separates fast, mocked tests from those that load real data over the
network, so CI (and developers) can run them independently. No test
logic changes: the same 34 tests pass before and after.

- Move tests/iddata/unit/test_load_data.py to
  tests/iddata/integration/test_loader.py. The old name referenced a
  load_data function that no longer exists after the 2.0.0 refactor.
- Extract test_hsa_populations into
  tests/iddata/integration/test_population.py. It exercises the
  PopulationData ancillary directly, not DiseaseDataLoader, so it does
  not belong in a file named after the loader.
- Add integration/conftest.py, which marks everything in that directory
  as integration. Note that pytest_collection_modifyitems is a global
  hook even when defined in a subdirectory conftest, so items are
  filtered by path.
- Register the integration marker in pyproject.toml to avoid
  PytestUnknownMarkWarning.
- Add empty __init__.py files to the test packages. Without them,
  pytest's default prepend import mode fails on duplicate basenames,
  which would break the first time a filename exists in both unit/ and
  integration/.
- Split the CI test step so network failures from S3/CDC/SEER/Census are
  distinguishable from real logic failures.

Refs #47
