+++ README.md
@@ -64,6 +64,19 @@ Developers will be using a Python virtual environment that:
     python -m pytest
     ```
 
+## Running tests
+
+The test suite is split into two kinds of tests:
+
+- `tests/iddata/unit/` — fast tests that use mocked data and make no network calls.
+- `tests/iddata/integration/` — end-to-end tests that load real data from S3, CDC, SEER, and census.gov. Every test in this directory is automatically marked `integration`, and the full set takes a few minutes to run.
+
+```bash
+python -m pytest                      # everything
+python -m pytest -m "not integration" # fast, offline tests only
+python -m pytest -m integration       # network-dependent tests only
+```
+
 ## Development workflow
 
 Because the package is installed in "editable" mode, you can run the code as though it were a normal Python package, while also
